[back] refactor: move the suggestion lib to the lib folder
