+++ backend/tournesol/lib/suggestions/strategies/__init__.py
@@ -1,0 +1,1 @@
+from .tocompare.classic import ClassicEntitySuggestionStrategy
