+++ backend/tournesol/lib/suggestions/strategies/base.py
