+++ backend/tournesol/views/suggestions/to_compare.py
@@ -3,10 +3,9 @@ from drf_spectacular.utils import OpenApiParameter, extend_schema, extend_schema
 from rest_framework import generics
 from rest_framework.permissions import IsAuthenticated
 
+from tournesol.lib.suggestions.strategies import ClassicEntitySuggestionStrategy
 from tournesol.serializers.suggestion import EntityToCompare
 from tournesol.views import PollScopedViewMixin
-
-from .strategies.classic import ClassicEntitySuggestionStrategy
 
 
 class ToCompareStrategy(models.TextChoices):
